requirements.txt
discord.py==2.4.0
PyNaCl==1.5.0
python-dotenv==1.0.1
yt-dlp==2024.08.06